# backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
    "/accounts": {
      "post": {
        "operationId": "CreateAccountController_handle",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Authentication"
        ]
      }
    },
    "/sessions": {
      "post": {
        "operationId": "AuthenticateController_handle",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Authentication"
        ]
      }
    },
    "/categories": {
      "get": {
        "operationId": "FetchRecentCategoriesController_handle",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Categories"
        ]
      },
      "post": {
        "operationId": "CreateCategoryController_handle",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Categories"
        ]
      }
    },
    "/foods": {
      "post": {
        "operationId": "CreateFoodController_handle",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Foods"
        ]
      },
      "get": {
        "operationId": "FetchRecentFoodsController_handle",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Foods"
        ]
      }
    },
    "/meals": {
      "post": {
        "operationId": "CreateMealController_handle",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Meals"
        ]
      },
      "get": {
        "operationId": "FetchRecentMealsController_handle",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Meals"
        ]
      }
    },
    "/mealItems": {
      "post": {
        "operationId": "CreateMealItemController_handle",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Meal Items"
        ]
      },
      "get": {
        "operationId": "FetchRecentMealItensController_handle",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Meal Items"
        ]
      }
    },
    "/wastes": {
      "post": {
        "operationId": "CreateWasteController_handle",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Wastes"
        ]
      },
      "get": {
        "operationId": "FetchRecentWastesController_handle",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Wastes"
        ]
      }
    },
    "/wastes/report": {
      "get": {
        "operationId": "GenerateWasteReportController_handle",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Wastes"
        ]
      }
    },
    "/categories/{id}": {
      "get": {
        "operationId": "GetByIdCategoryController_handle",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Categories"
        ]
      },
      "put": {
        "operationId": "EditCategoryController_handle",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Categories"
        ]
      }
    },
    "/foods/{id}": {
      "get": {
        "operationId": "GetByIdFoodController_handle",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Foods"
        ]
      },
      "put": {
        "operationId": "EditFoodController_handle",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Foods"
        ]
      }
    },
    "/meals/{id}": {
      "get": {
        "operationId": "GetByIdMealController_handle",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Meals"
        ]
      },
      "put": {
        "operationId": "EditMealController_handle",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Meals"
        ]
      }
    },
    "/mealItems/{id}": {
      "get": {
        "operationId": "GetByIdMealItemController_handle",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Meal Items"
        ]
      },
      "put": {
        "operationId": "EditMealItemController_handle",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Meal Items"
        ]
      }
    },
    "/wastes/{id}": {
      "get": {
        "operationId": "GetByIdWasteController_handle",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Wastes"
        ]
      },
      "put": {
        "operationId": "EditWasteController_handle",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Wastes"
        ]
      }
    },
    "/categories/{categoryId}": {
      "delete": {
        "operationId": "DeleteCategoryController_handle",
        "parameters": [
          {
            "name": "categoryId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Categories"
        ]
      }
    },
    "/foods/{foodId}": {
      "delete": {
        "operationId": "DeleteFoodController_handle",
        "parameters": [
          {
            "name": "foodId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Foods"
        ]
      }
    },
    "/meals/{mealId}": {
      "delete": {
        "operationId": "DeleteMealController_handle",
        "parameters": [
          {
            "name": "mealId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Meals"
        ]
      }
    },
    "/mealItems/{mealItemId}": {
      "delete": {
        "operationId": "DeleteMealItemController_handle",
        "parameters": [
          {
            "name": "mealItemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Meal Items"
        ]
      }
    },
    "/wastes/{wasteId}": {
      "delete": {
        "operationId": "DeleteWasteController_handle",
        "parameters": [
          {
            "name": "wasteId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Wastes"
        ]
      }
    },
    "/meals/{mealId}/items": {
      "get": {
        "operationId": "FetchMealItemsByMealIdController_handle",
        "parameters": [
          {
            "name": "mealId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Meal Items"
        ]
      }
    }
  },
  "info": {
    "title": "Waste Control API",
    "description": "API for controlling food waste",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {}
  }
}